Move Automap instance code to a new class named Automap_Map
As the instance code is slow path is extension, moving it to a separate class will allow to remove it from the extension.
Automap API changes: rename instance() to map(), new base_path() method, success handler receives an ID instead of a map object (this was breaking the fast path).
Automap_Tools::check() receives an ID instead of a map object, as the map object does not know its base path anymore.